requirements.txt
streamlit
dspy-ai